Move MetricRecorder to OpenTelemetry
