+++ dev/com.ibm.ws.microprofile.faulttolerance.telemetry.metrics.2.0/bnd.bnd
@@ -13,22 +13,25 @@
 -include= ~../cnf/resources/bnd/bundle.props
 bVersion=1.0
 
+openTelemetryVersion=1.35.0
 
-Bundle-Name: com.ibm.ws.microprofile.faulttolerance.metrics.2.0
-Bundle-SymbolicName: com.ibm.ws.microprofile.faulttolerance.metrics.2.0
+Bundle-Name: com.ibm.ws.microprofile.faulttolerance.telemetry.metrics.2.0
+Bundle-SymbolicName: com.ibm.ws.microprofile.faulttolerance.telemetry.metrics.2.0
 
-Bundle-Description:MicroProfile Fault Tolerance metrics integration
+Bundle-Description:MicroProfile Fault Tolerance telemetry integration
 
 WS-TraceGroup: FAULTTOLERANCE
 
 -dsannotations-inherit: true
--dsannotations: com.ibm.ws.microprofile.faulttolerance.metrics_20.MetricRecorderProviderImpl
+-dsannotations: com.ibm.ws.microprofile.faulttolerance.telemetry.metrics_20.MetricRecorderProviderImpl
 
-Import-Package: *
+#We need an explicit version for mp config because we're not transforming this bundle.
+Import-Package: org.eclipse.microprofile.config;version="[3.0,4)", \
+        *
 
 Export-Package: 
 
-Private-Package: com.ibm.ws.microprofile.faulttolerance.metrics_20
+Private-Package: com.ibm.ws.microprofile.faulttolerance.telemetry.metrics_20
 
 
 -buildpath: \
@@ -45,7 +48,9 @@ Private-Package: com.ibm.ws.microprofile.faulttolerance.metrics_20
         com.ibm.websphere.org.eclipse.microprofile.config.1.3;version=latest,\
         com.ibm.ws.org.osgi.annotation.versioning;version=latest, \
         com.ibm.ws.microprofile.metrics;version=latest,\
-        com.ibm.ws.microprofile.faulttolerance.metrics;version=latest
+        com.ibm.ws.microprofile.faulttolerance.telemetry;version=latest, \
+        io.opentelemetry:opentelemetry-api;version='${openTelemetryVersion}';strategy=exact, \
+        io.openliberty.microprofile.telemetry.internal.common
 
 -testpath: \
 	../build.sharedResources/lib/junit/old/junit.jar;version=file, \
