+++ dev/com.ibm.ws.microprofile.faulttolerance.telemetry.metrics.2.0/src/com/ibm/ws/microprofile/faulttolerance/telemetry/metrics_20/package-info.java
@@ -11,4 +11,4 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 @com.ibm.websphere.ras.annotation.TraceOptions(traceGroup = "FAULTTOLERANCE")
-package com.ibm.ws.microprofile.faulttolerance.metrics_20;
+package com.ibm.ws.microprofile.faulttolerance.telemetry.metrics_20;
